api-gateway/requirements.txt
fastapi
uvicorn
httpx
PyJWT
prometheus-client
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
